ding: Use fully async directories enumeration via FileUtils
While we were already using async enumeration of folders we didn't query
file infos asynchronously, solve this using an async files enumerator
and using promised asyc queries everywhere.

Fix also the case in which we could have tried to re-use a newly
recreated cancellable on subsequent calls, ensuring that we only use a
local cancellable in all the async calls

As per this we can just mark Gio.File.enumerate_children_async as a
promise all the times.
Fixes

Continuation