dbusUtils: Add ability to push events to RemoteFileOperationsManager
With async functions we may have the need to use events that are not
available at the moment we call the RemoteFileOperationsManager methods,
so we need to pass it to the methods when they're called.

Not to change the signature to all the methods in both legacy and
current implementation, it's just better to define it via a stack of
events that can be consumed at each call.