Get Intelligent menu position from the grid depending on where clicked.
Get Intelligent menu position from the grid depending on where clicked.

Currently if dock is hidden in intellihide mode, menus can go under the dock.

When Dock or other objects are hidden in intellihide mode, they set usable area margins. This detects these margins in visibleArea.js and passes it to the grid. Menu click position is then used to determine the side on which the menu should be shown to prevent the menus from going under the dock or other hidden objects on the desktop.

Fix desktopBackground menu as well as askrename popupmenu so they do not go under dock when it is in intellihide mode.

Fix Rename Popup

Fixes