Revert metadatachanged to update entire desktop
When a file is renamed, the position can jump on the desktop. This appears to be a race condition where the metadata::saved-coordinates are read as null by readfilelist which is very fast, while the metadata-updated has not refreshed with the saved-coordinates of the new file. This forces a desktop re-read, resulting with icons always in the correct position on rename.