When reassigning files to desktop, saved position is set to null in desktopManager, and then finally assigned a position when added to desktop by desktopgrid.
This results in two writes of the saved position immediately after one another. As saving is asynchronus
now, the second write clobbers the first and cancels it. A side effect of this is that the saved coordinate
is reverted to the old coordinate in local memory, however  the new coordinates are written to disk, they are no
longer in sync.

This results in discrepency where icons jump around to incorrect earlier places when moving in and out of stack
mode or arranged icons mode.

This commit introduces a temporary position that is held in memory, assigned in desktopManager during drag
and drop as well as when reassigning icons to the desktop. The final assigned coordinates are then written to
disk only once during the final phase in desktopgrid when added to the grid.

The coordinates are written asynchronously only once and there is no cancellation, making the program more
efficient and fast.