Fix Unlimited Workspaces with auto-move-windows
auto-move-window replaces _checkWorkspaces() with it's own custom function that causes unlimited workspaces to the right when DING is enabled. This is because the function detects the DING windows and marks the workspace occupied and used even when there is only the DING window on the desktop.

This monitors auto-move-windows and overrides the gnome shell with our custom methods.
Fix Error