Remove move to active workspace
As the DING window is stuck() now, it is on all workspaces. There is no need to move the window to the active workspace, remove that code.

Move the window to the bottom as soon as the workspace is switched, as we do not need to move the window to the workspace.

This takes away the need for the moving window workspace animation hack delay, to hide flickering icons. Remove that shell override as well

Further cleanup, remove code to move window to active workspace

Rename function to make more sense.