Fix 'stickyness' loss
The DING window actor can be dragged and dropped in other extensions that show a thumbnail of the actor. This results in unstick() as well as moves the window to a particular workspace.

This insures that the stick() property is restored, and further checks to make sure that the window is on the active workspace.