this._windowlist needs to be initialized properly and not null in the constructor.
When enable() and disable() are called rapidly on the lock screen, disable() can be called before this,_windowList is initialized
in emulateX11.enable(). The disable() call calls emulatedX11.disable(), that does not expect this._windowList to be null.
If it is null the extension disable() call fails resulting in generalized mayhem and multiple DING processes
being launched.

Thanks to extensioons matrix chat channel for help in coming up with this fix.