Nautilus shows rectangular icons by default. This changes the style of the icons to rectangular.
Although this is similar to Nautilus, it is not the same. The size of the highlight fits
the height width of the icon, still leaving space between the icons to allow
drag and drop to those areas of the grid.

The sytle can be toggled and is linked to highlight drop target