Keep distance between icon and label constant with icon shape change
With change in shape of icon from trapezoidal to rectangular, the distance
between the icon and label changes.

The icon containger Gtk.Box has a padding applied in css. So does the label
container.

With rectangular icons we are not putting the label in a box, so there is no
padding around the label, consequently the label appears closer to the icon.

This MR compensates for the distance by applying a top margin to the label
in css on shape change to keep a consant distance from the icon container.