Use Cairo, poppler to draw a pdf thumbnail icon.
We can make thumbnails of common image formats using gdkpixbuf.

This is used as a fallback when gnome-desktop thumbnail factory is unable to render a thumbnail.

Import poppler and Cairo to create thumbnails for  pdf files as well if thumbnail-factory is unable to create thumbnails for pdf files.

Use Cairo, poppler to draw a pdf thumbnail icon.