Fix thumbnail icon margin
Generic icon picture items have a margin of 4 around their default size.

Thumbnails generated by gnome desktop library and our custom icons have no margins

The placement of the icon from the top of the icon and from the bottom text therefore
differs. The generated thumbnails are 4 pixels nearer to the top, and to the bottom label text.

Add a margin to generated thumbnails to make everything even.