Stop open folder on drag hovering if drop complete
When spring loaded folders on drag hover are enabled, the folder with the drop opens even when the drop is completed before opening of the folder from the hover timer.

In this case, the user did not want the folder open for further navigation, and this is not necessary.

So if the drop is complete and successful prior to auto hover open of folder, stop all timers and opening of the folder.