Leave the dragged item highlighted
Currently when a fileItem icon is dragged, it looses its highlighted status with drag motion and updates.

If there are multiple items in the selection being dragged, the other icons remain highlighted, except the dragged icon.

Leave the dragged item highlighted during drags with this change, consistent with the other icons.