Update activeFileItem on deletion
If an activeFileItem is deleted or moved off the desktop, a reference to the object and file is held.

If at this point, the enter key is pressed, the program tries to open a non existent file from the object via shortcut actions.

Although the program does not crash, and fails gracefully, it writes ugly messages in the logs.

After a desktop refresh and update, if the activeFileItem does not exist, therefore set the value to null so the open code can no longer be called.