Rewrite refresh Trash Icon
Simplify the code to refresh the trash icon.

Do not use a refresh timeout with GLib.time_out_add(). We can throttle the number of updates to once a second. Therefore there should be no need to cancel prior updates, a cancellable is not really needed.

Do, not use a trash Gio.cancellable, it is not needed. An automatic queryFileInfoCancellable is created by the underlying code.

There is similar code in updateSymlinkIcon, refactor to put into Icon Rendering section.

Use name constants.

queue_draw the icons after update to trigger appearance updates, On the newer releases and some distributions, the icons were not getting updated.