updateIcon global is used in desktopManager to only update the icons, not read the metadata again and redisplay the icons.
It is used in the base class, desktopIconItem.

Rename the newly writtern updateIcon in fileItem, which is overriding the base class to _reloadIcon so that it does not override the base class and
allows use without re-reading the metadata with the corresponding async read of file.

Use the _reloadIcon() to actually re-read the file, reload the metadata, and then redisplay the icon.

This prevents double read of file through desktopManager.

use local _reloadIcon() in fileItem.js to redisplay the icon on changes.