Move files from trash instead of copy
Files in trash have a filesystem id of 'trash:' instead of the
the actual ID of the filesystem containing the trash.

Therefore, when dropped onto the desktop they appear to be from
a different filesystem and are copied instead of being moved to
the desktop.

If filesystem id is 'trash', consider it being from the same
fileystem as the desktop folder.