Open appimage files properly
Write code to open appImage files properly.
If there is an app registered to integrate appImage files, it is launched.
If there is no app, check and see if file is executable and trusted.
Give appropriate notifications if the appImage file is not executable.
Give appropriate notifications if the appImage file is not trusted.

If the appImage file is trusted, it will be launched.