Add allow launching for appImage files
Show Allow/UnAllow launching for app image files in the right click menu.

Execute code to set metadata:trusted value and set executable from right click menu.

Do not allow the file to be launched as a Program from the right click menu, remove that option. That is reserved for scripts.