Always return updated name of the current set Desktop
GLib.get_user_special_dir() returns the desktop file set at program start.

If the desktop file is changed with xdg utilities, this is not reflected
in the GLib call till the app is restarted.

Reading the .cofig/user-dirs.users file will always give the correct
currently set desktop with the xdg utilities.

So always return the Desktop folder set in .config/user-dirs.dirs if
set and available.

If not available, fall back to the GLib call. This will however only
return the folder set at program start.

The GLib call falls back to $HOME/Desktop if it cannot read the xdg
set desktop.