Handle incorrect desktop Folder
xdg desktop may be set incorrectly by the user. It may point to a
file, or a non existent folder, or pipe etc.

Build in sanity checks to make sure that is a folder before
switching to the new folder.

Also perform sanity checks at startup to make sure we have a
proper folder set as the display desktop. If this is not so,
throw up an error message.

If the folder is incorrect, there are multiple errors logged by
doReadAsync trying to read a non existant directory.

Build in try catch and other codes to catch this condtion and bail
out gracefully. Just log a message that the desktop is not set.

Do that for updateWritableByOthers() as well to handle
Gio.IOErrorNum.NOT_FOUND