Give visual indication on desktop window that desktop folder is not set correctly
If there is no desktop folder, make the desktop red.

Fix errorstate

Only run after the error dialog is run.

Running prior to that has a race condition, the grid windows are not
yet mapped and or for some reason the errorstate is not set.

If we wait till the error is acknowleged, everything works perfectly and
the window background changes to red.