Give Notification of Desktop Changes and Aborted Changes
Add notification if the new desktop folder does not exist and the change is aborted.

Add notification on switching to a new desktop.

Call the function as a timeout, to prevent multiple activation's, as the changed signal may fire many times for one change.

Try and minimize activation's for the same change, by only changing on the Gio.FileMonitorEvent.CHANGES_DONE_HINT

xdg-user-dirs-update 'renames' the file, there is not changes-done-hint

Update the code to take care of this eventuality