Add application name and some default icons to notifications
The appname is default set to 'Desktop Icons'

The icon displayed is 'computer' which should be part of freedesktop icon
them specification and should be available.

The image displayed is computer.svg from the Adwaita theme scalable
devices. The Adwaita theme should be installed on all Gnome Desktops
and should be available.

Your mileage my vary.

However, this is the simplest and quickest way to display icons, without
installing a .desktop file and using Gio.Notification

This can be tweaked in the future if we install Gtk4-Ding as a proper
application with a .desktop file.

Reformat and use constants to make meaning clear