Sync Settings and the radio Menu
Add a sync function

Changing preferences arrangeorder will activate the action to the new change

Changing the radiomenu will activate the action and change the settings to the new settings.

If nothing is changed, the desktop will still be refreshed to the currently set arrangeorder

Do not sort the icons twice.

Setting the desktopSetting for arrange order on  radio button menu clicks calls the sync function recursively, so it is called twice with the same value.

This sorts the icons twice on the desktop.

We cannot block the signal handler for the desktopSettings to prevent this.

Build in logic to the sync function so that sorting the icons happens exactly once even when called recursively.

Block the action on the second call as we cannot block the signal