Remove textaccelerators from show error.
Previous dialogs would activate the application text shortcuts when errors
were showing instead of the local accelerator for the dialog.

This led to the workaround of turning them off when showing dialogs
and then turning them on when the dialog was dismissed.

The new Adw.AlertDialog does not do that any more.

Therefore remove use of these accelerators.

Also just pass the timout function, instead of entire desktopIconUtilities.

Also, get application from Gio call instead of the desktopIconUtilities.