Previously, there was a regression and window.comput_point would return the incorrect coordinates after grid resizing.
The workaround was to use old methods to copute co-ords.

This let to error in the test window where the header bar would throw
off the coordinates by the height of the header bar.

This seems to be fixed in the latest Gtk4/Adw.

Revert to prior more elegant way of computing relative
co-ordinates, using Gtk/GObject calls.