Use Adw.StyleManager
Use Adw.StyleManager to monitor for all changes to accent/background colors, darkmode.

Use Adw.StyleManager to set the accent and background colors.

Init and monitor all color changes in Preferences.js, as we monitor all preferences and changes there.

This reduces the size and complexity of code in desktopManager.js