Save normalized coordinates with the monitor index for each file Item
Extend and save nautilus-icon-position as string representing a 5 element array.

The first two are the global coordinates on the screen.

The next two represent normalized x and y values for the container

normalizedX = localX / this._width;
normalizedY = localY / this._height;

The width and height here are the width height of the local container and correspond to the local coordinates. At this level, all local coordinates, width and height are derived after correction for margins, zoom/sizer

The last represents the monitor index the icon is placed on.

If the monitor resolution changes, or the primary desktop is changed, we can then scale the icon to the correct relative position on the new monitor or resolution by just multiplying the normalized co-ordinates by the width and height respectively of the new container. 

The new container width and height are derived after applying the new margins, zoom/scale, to make it independent of the margin, zoom/scale.

The new global co-ordinates can then be easily calculated for the new monitors or resolution.

The nautilus-icon-position can also be saved with two elements, representing the global co-ordinates, and can also be parsed as a two element array representing the global coordinates to make it compatible with the old program and other programs that may save data in the same string.

Clean up writing and saving coordinates with simpler and cleaner code.