Recompute new co-ordinates
When the monitors change in resolution/size, number (addition/deletion), recompute the global coordinates of each file Item for the new monitor, resolution using the normalized coordinates.

The positions are remapped to the new monitor, or new resolution. The new monitor may have a different resolution, zoom, margin as well - all this will be taken into account.

Recompute grid layout positions if  only grid changed, not the monitors or resolutions.

This prevents unnecessary 'jumping' of icons to random new positions, they are simply scaled remapped to the same relative positions.

Improve regridding