Fix remote activation via command line
Remote invocation of command line did not work on the running application.

The problem - pushing the parsed parameters into this.desktops and this.desktopVariant arrays just added to the values already present in the array in the remote running application.

Initialize the array and fill them with the new parsed values, then process them remotely.

Also add updateGridWindows action to the main.app directly, so it can be activated without getting the remote ding actions.