Fix remote activation
When the app is activated remotely through command line, there were various errors, where the running app variables were affected.

This would just make the array of desktops grow with the old one, instead of replacing with the new one.

This simplifies logic and code, running desktops are just updated with the new one from the remote command line. This allows resizing, redoing the grid and margins from the command line, specially for testing purposes.

The other variables are updated from the command line, but do not have any affect as they are not updated in the running instances of the classes.