Fix monitor resolution changes with stacks enabled.
monitor size changes in a list of monitors, while stacking is enabled, refresh the saved coordinates of the stacked items to the new geometry of the monitor.

This ensures that when stacking is disabled, the icons will return to same relative positions on the new monitor resolution/area.

When the number of monitors changes, stack is not redisplayed, the actual fileList read from disk with the initial saved coordinates, and is restacked from scratch again. So that is a different problem-

Transforming coordinates at that point needs to happen when actually creating new grid windows.