Create a stack of mime-type icons to represent the stack top icon
The stack top renders as a generic folder icon.

When the stack is open it renders as a folder open icon.

Folder open icon is not installed on some themes.

Instead of adding both icons to Gresources as local asset, render the stack top icon distinctively as a stack of mime-type-icons for that stack.

This is instantly recognizable and it becomes easy to distinguish one stack type from another stack type.