Initialize properly on startup
Startup by insuring that the grid windows are created.

Then show any errors, as now we can actually display them.

Then updated the desktop, mark as initialRead.

Also mark as initialRead if a new desktop directory is set.

if initialRead is set, then all fileItem positions will be recomputed to the new monitors geometry - we can actually insure that this can be done as windows are created and geometry is available.

If monitors are changed, then recompute fileItem positions as well.

Windows created can be a local constant, does not have to be
global, and thus does not have to be destroyed.