Fix window coordinates for multi-monitor setups
On multiple monitor setups, on the second monitor, or any monitor to the right, the co-ordinates are calculated incorrectly, and offset to the right.
This is because, when reverting to this._window.compute_point(wiget, point), converting the point to window coordinates step was erroneously left out.
Fix this problem.
Fix for #127