Monitor placement algorithms
This allows us to place the icon on any monitor we want.

It takes into account, first the prior monitor the icon was placed.

If it is the primary monitor, it is placed on the new primary monitor.

Unless place on secondary monitor is preferred, where it will be place on a secondary monitor instead.

If the icon is not on a primary monitor, and the same monitor is available in the new setup, it is remapped to the same monitor.

If the prior monitor is not available, it is mapped using placement algorithms, respecting the users choice for placing on primary monitor or secondary monitors in preferences.

Needs addition of geometrical placement algorithms on monitors, possibly depending on start corner of icon arrangement in future.