Get and set fractional scaling in preferences
Keep updated with change in mutter settings.

Remove un-needed code in desktopManager, use getters and setters
in desktopManager to get/set the properties from preferences
Show error dialog and notification if fractional scaling is not enabled

In multiple monitor systems, if they have different zooms/scales, fractional
scaling is necessary to display the windows with the correct dimensions.

Detect if windows have different zooms.

Detect if fractional scaling is enabled in mutter.

If there are different zooms, and scaling is not enabled, after
windows are created, display an error, stop.

Also show a notification as the windows may be incorrectly sized
and the error dialog may be off center.

Set desktop scaling correctly