Update desktop every read
When the program starts, there are some triggers that seem to trigger multiple updateDesktop() with opts.initalread = false. This results in random placement of some icons on startup as the the initial placement has not yet finished and the new icon position has not been written to metadata by initalread=true.