Fixes to recompute window positoions for any monitor or margin or grid change
Recompute new grids if snap to grid to prevent icon migration

When margins are changed by dock or other actors on the side of the screen in any direction, the container migrates and expands to fill the empty space.

The grid is recreated, and the icons systematically fall into new grids, depending on where they land, with bias towards the top left corner when snap to grid is selected.

This is specially true for opening and closing of the laptop lid as the dock extension is disabled and enabled, making the margin disappear and come back again and again, and the icons progressively keep migrating in one direction/column/row depending on the dock position with every open close of the lid..

This adds a 'shift' bias to every icon to move in the opposite direction so that they end up in roughly the same position, as now they will tend to fall to the grid on the other side of the shift.

This is true for margin changes approximately the width/height of icons, however, may not be true of very large margin changes.

However in current working conditions, margin actors are about the same size as the icons and this works nicely in practice, with of course some exceptions where the entire column may shift to the left or right leaving empty grids between icons when clustered together. The opposite happens when the margins come back, so in the end the icons do not progressively shift in one direction.

Set temporary monitor index for grid recomputation if snap to grid is selected.