Move geometry computation to desktopGrid
It is only done once on grid changes when the grids are created. This prevents calculating shifts in a loop for every fileItem.

Re gridding is done after grids are created, and these don't
have to be calculated in a loop again and again.
Prevent shift up into another monitor- ShiftUp sign was wrong.