Create DesktopMonitor Class
Move all code that monitors the desktop folder, changes the desktop folder, to its own class.
This simplifies DesktopManager class, and removes a huge chunk of code to its own class.
This new class simply maintains an updated fileList with all the icons of the desktop all the time, and just signals the main class to redraw the desktop when there are changes.
This DesktopMonitor fileList is now different from the desktopManager fileList. Since the Manager draws the desktop, its fileList is now named displayList.  They are synced just before the the desktop is drawn.
The desktop render and draw can thus be finely controlled from the manager. The manager does not allow the desktopMonitor class to update the desktop unless a display is ready and available. The desktop is drawn only when a monitor/window is available, and at that time all coordinates are rewritten using saved normalized coordinates and monitor information. This insures that all icons retain monitor and relative positional information all the time.