Add methods to refresh, redraw , reload and reframe the desktop
Remove _updateDesktop

All desktop changes are made through-

refreshDesktop- get the latest fileList from desktopMonitor and redraw desktop
This alwasy uses the normalized coordinates from the newly read fileList to
recompute window and grid positions for each fileItem.

reDrawDesktop- use current fileList in desktopManager and redraw desktop,
positions are not recalculated, only rendering of fileItems etc are changed.
For theme changes etc.

reFrameDesktop- only used thrwough windowManager - when geometry changes,
uses the current fileList in desktopManager, recomputes new coordinates
and grid coordinates, and redraw the desktop with the new geometry and
positions.

reLoadDesktop forces a new read of the fileList, that then calls refreshDesktop to draw a new
desktop.

Add method clearAllFilesFromGrid that removes the fileItems from grid without destroying them,
so that their UI can be updated and then they can be reassigned to grid.

This is in contrast to removeAllFilesFromGrid that actually destroys the fileItems so that new ones are created.

Drawing is only allowed after windows are available to calculate new
positions, these are immediately written to disc when the icons are
rendered and placed in the new window.

We _syncStart the desktop on initiation, to insure all positions
are calculated properly on startup - startup only occurs when
the windows are created and available to display the icons. Drawing
the desktop is not allowed, is blocked with a promise that is only
resolved when windows are created.

Then desktopMonitoring and preferences monitoring takes over and
refreshes/redraws desktop as necessary.