Create DesktopBackgroundMenu and DesktopActions Classes - Initial seperation
Creates desktopMenu.js file that has two classes-


desktopActions class moves all code for actions associated with the background menu to it's own class and simplifies desktopManager. It adds all the actions to the main application and has all the methods for executing the actions.


desktopBackgroundMenu class abstracts all code for making and displaying the right click menu of the desktop into its own class and moves it away from desktopManager.


This simplifies the manager and makes further maintenance simpler.

Add a desktopMenu.js file