Initial refactor
Create a copy of drag and drop functions in a new class in a separate file.

Moving pendingSelfCopyFiles and pendingDropFile objects to this new class.

Use getters and setters to refer to these objects in desktopManager and all other classes.

Refer to gnomeshelldragdrop in desktopManager and pass it to the new class. This avoids having to import the module in the new class, and keeps all module imports
centralized in loclafiles.js.