Fix trash Undo
We now save a five position array for position of icons.

When the item is trashed, the coordinates are nulled. Therefore, for undo, we have been saving the coordinates in pendingSelfCopyFiles object. However we are only saving the global coordinates!

When the file is restored, desktopMonitor assigns the 2 position savedCoordinates from pendingSelfCopyFiles. However when resetting the position to account for zoom, in recomputewindow positions, since there are no normal coordinates or monitor info, the icon is set as a non assigned icon and ends up on an available empty space.

The complete solution would be to save the 5 position array in pendingSelfCopyFiles, and then reassign that on undo. This accomplishes that.

Save 5 position coordinates in pendingSelfCopyFiles, these are then restored when read from disk, so the Icon ends up in exactly the same place.