Create three new classes for DesktopGrid
Transition DesktopGrid to three classes-

Base class DisplayGrid that manages all the geometry and icon placement.

Extend the class to DrawGrid, that add functions for drawing on the desktop- like rubber
bands, highlighting and drop rectangles.

Extend this class yet again to add drag and drop support, and this becomes the final
DesktopGrid, that is called as usual from the desktopManager.

This simplifies maintained code and reduces clutter.

Also format the entire file with extra white space and less than 80 char length
for legibility and readability