Split FileItem into different classes
This first commit simply refactors fileItem.js to 80 characters for readability and further transformation.

FileItem is getting quite complicated and long.
The objective of this MR, through a long series of commits, is to transition and break up FileItem class into smaller separate classes, encapsulation methods for that special file type, to make the overall logic and flow simpler to understand.
SpecialFolderIcon class handles trash and the home directory on the desktop
VolumeIcon class handles all mounted volumes on the desktop
DesktopFileIcon class handles .desktop files on the desktop
AppImageFileIcon class handles all methods for appimage files on the desktop
All the above classes extends a new simplified FileItemIcon class that replaces FileItem class. This in turn extends the basic desktopIconItem class.
SymLinkIcon class handles symlinks. It extends the base class it points to.
They are all called through a new IconFactory class that picks the right class to use for the file in question.
This should make it easier to maintain the code in future, and to easily extend it for specific purposes, for special icons etc.