Move de-compressible types to DBusUtils
Currently we monitor for decompressible types of files in menuManager.

Simplify menuManager and fileItemMenu, move monitoring for the decompressible types to DBusUtils remote file operations, right next to the proxy it is querying. This is a more logical place for it. It keeps an updated list of types supported as a property array.

Just get a list from RemoteFileOperations when needed, while constructing the menu.