Add a shortcuts window
Create a GtkBox with a header bar and a scrolled window with a listbox that shows all the current actions which have keyboard accelerators, and show all the keyboard shortcuts.

Add a state hint to every action with an accelerator, hijack this information to set as the description in the Gtk.Box for the keyboard shortcut..

Show this box with a shortcut in a new Adw.ApplicationWindow.