Add modal dialogs to managed windows
The shortcuts window is mapped as a modal dialog, integer enum 4.

It is therefore not managed by emulateX11, and appears in the
window list, taskbar. It therefor also does not stay on top
and can be lost beneath other windows. As it is modal for the
destop window, if lost, user cannot interact with the desktop.

Change the exclusion criteria so that we can manage windows of
this type, hide it from the taskbar and keep on top.