Connect to unmanaging signal rather than unmanaged
This signal is emmitted prior to window being destroyed,
and we can disconnect all our handlers in customJS_ding
prior to the window being destroyed and unmanaged.

When connecting to the umanaged call, the window is
null in customJS_ding, and if the window is handled,
the handlers in that class generate error messages in
gnome shell, prior to being completely disconnected.