Add a .desktop File and app Icon to the users local folders
On app startup, a .desktop file for the app is installed in
~/.local/share/applications/

On app startup, an icon file is installed to-
~/.local/share/icons/hicolor/scalable/apps/

Catch error if parent folders do not exist, in that case create them.

Both the icon and the .desktop  file are stored in the /data
GResource known location.

Check the user installed files or if files already exist, in which case
update them only if different. GLib.memcmp() does not exist,
write our own function to do memcmp for this purpose.

Install the icon and .desktop file only if running as a local extension.

A system install does not require us to place an icon or a .desktop file,
as they can be installed by the system install scripts or package manager
Consequently, the files are only removed from the correct folders when
running through a local extension

All this is done asynchronously to not delay startup.

The application thus knows if a write was successful, it can then-
a. then only remove files if we actually were the ones writing the file.
b. Update the icon cache only if a new icon file was written.
c. Update the desktop-database only if a new .desktop file was written.
d. Update the icon cache before the desktop-database

This will not remove files if the user installed them and we did not
write the file on startup ourselves, or if we did not need to update
an older file.

Finally, log, if local install, and if successful,  file installation and removal