Make updateMetadataFromFileInfo synchronous
Remove the await from setEncryptionStatus(), this updates and
reloads the encrypted emblem in the background, while making
updateMetaDataFromFileInfo synchronous.

setEncryption status is async in bacground as it has to await
checking encryption. If encrypted, the icon will be reloaded
with the correct emblem in the background.

This allows us to updateName properly after metadata is
updated in a synchronous way in the constructor for
fileItemIcon.

Remove async await for updateMetadataFromFileInfo() in all
extended classes as it is synchronously called in the superclass
in the constructor.