Subclass Adw.ActionRow to make display and editable shortcut rows
DisplayShortcutRow subclass displays the current and default shortcut.
EditShortcutRow subclasses DisplayShortcutRow to allow editing the shortcut.
The default parameters are the action name, current application, ie the actionmap,
the callback to read the shortcut from settings, the callback to write the new shortcut to settings.
Add clickable icons to the editable text entry to reset accelerator to default or to disable the accelerator completely.
The text/key entry dialog by design only allows new shortcuts with a modifier key, one key shortcuts are not allowed, as there may to to many clashes with global shortcuts.

The text/key entry dialog by design only allows new shortcuts with a modifier key, one key shortcuts are not allowed, as there may to to many clashes with global shortcuts.