Show shortcutwindow through preferences window.
Show the normal editable shortcutwindow, triggered
via remote actions, on clicking a button in preferences. This
reuses the code for the shortcutwindow without writing seperate
code for preferences.

Shortcutwindow only works correctly when the app is enabled.

When the preferences window is called through gnome extensions, the app
may not be running if the extension is not enabled. Connect to the
DBus.watch_name() for the app in this case to get remote actions.

Trigger the window through the remote action in this case. The button
sensitivity is tied to the actionGroup being available.