Add a .desktopFile actions section to the right click menu
if the .desktop file is trusted and has actions this menu is retrieved and added to the
right click menu. This should list all the actions from the .desktop file, with the associated action. This is added to fileItemMenu.

Add an action to the main application to execute the .desktop file actions. It accepts
a Gvariant parameter that is an array of the .desktop file path, the action name and the actual action string in the .desktop file. This is added to the the fileItemActions class.

Add a method to execute the action- checks if the active file Item is the actual .desktop file by comparing paths, checks to see if the actionMap for the activeFileItem has the action. If both these pass, then it retrieves the associated desktopAppInfo property of the fileItem to finally execute the the chosen action.