Move all methods to manipulate and query the user-dir.dirs file.
Create a class with all the methods to manipulate user-dir.dirs and and keep track of the currently set desktop folder.

This is to finally remove all dependencies on xdg-user-dirs-update and xdg-user-dir. The same code can then be used in adwPreferencesWindow and desktopFolderMonitor without duplicating the code to monitor and manipulate the file.

We thus only use GLib and Gio and introduce no other dependencies on external programs.