Promisify Gio.File.replace_contents_async
When DesktopFileUtils is called through the gnome shell, using the
extnsions prefs.js, replace_contents_async does not work as coded
as it has not been promisified.

When called through the application, it works as Adw.Ding.js
does the promisification already.

Promisify this Gio call when called through prefs.js directly
from the gnome shell and extension settings or extension-manager.