Set locale aware name for the Desktop Folder
When restoring or checking the default desktop, use the localized name
for 'Desktop' like xdg-user-dir does, translating the name.

First look for /etc/xdg/user-dirs.defaults file. Find the string used for 'DESKTOP'.

Use gettext to lookup the translation for the DESKTOP string found above in the 
xdg-user-dirs domain.

If there is no user-dirs.defaults file available, or it does not have a string set for
'DESKTOP', default to using Enums.DEFAULT_DESKTOP_NAME, which is currently
set to 'Desktop' in English. This can be changed by the user in Enums.js

If translation is found, use the translated string or default to using Enums.DEFAULT_DESKTOP_NAME.

Check if this directory exists in the $HOME folder, if not, create it.

Set the desktop to the new localized folder.
Force write the desktop folder if it does not exist